+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/ProcessCompilerData.scala
@@ -9,7 +9,7 @@ import pl.touk.nussknacker.engine.compile.nodecompilation.NodeCompiler
 import pl.touk.nussknacker.engine.compiledgraph.CompiledProcessParts
 import pl.touk.nussknacker.engine.definition.DefinitionExtractor.ObjectWithMethodDef
 import pl.touk.nussknacker.engine.definition.ProcessDefinitionExtractor.ProcessDefinition
-import pl.touk.nussknacker.engine.definition.{LazyInterpreterDependencies, ProcessDefinitionExtractor, SubprocessDefinitionExtractor}
+import pl.touk.nussknacker.engine.definition.{LazyInterpreterDependencies, ProcessDefinitionExtractor, SubprocessComponentDefinitionExtractor}
 import pl.touk.nussknacker.engine.dict.DictServicesFactoryLoader
 import pl.touk.nussknacker.engine.expression.ExpressionEvaluator
 import pl.touk.nussknacker.engine.graph.node.{NodeData, WithComponent}
@@ -29,7 +29,7 @@ object ProcessCompilerData {
 
   def prepare(process: CanonicalProcess,
               definitions: ProcessDefinition[ObjectWithMethodDef],
-              subprocessDefinitionExtractor: SubprocessDefinitionExtractor,
+              subprocessDefinitionExtractor: SubprocessComponentDefinitionExtractor,
               listeners: Seq[ProcessListener],
               userCodeClassLoader: ClassLoader,
               resultsCollector: ResultCollector,